// src/types/FormData.ts
export interface ApplicationFormData {
  firstName: string;
  lastName: string;
  dob: string;
  snapBenefits: string;
  snapPrimaryRecipient: string;
  snapRecipientName: string;
  snapBenefitCity: string;
  snapBenefitState: string;
  tanfBenefits: string;
  tanfNineMonths: string;
  tanfRecipientName: string;
  tanfBenefitCity: string;
  tanfBenefitState: string;
  ssiBenefits: string;
  unemploymentBenefits: string;
  unemploymentState: string;
  signature_url: string;
  formStartTime: string;
}